Fixed the animation to sub sub bubbles. AddVC still saves even when stuff isnt entered yet though. Name on setup doesnt save either

+++ NCEACredits/AnimationManager.h
@@ -15,7 +15,7 @@
 @interface AnimationManager : NSObject
 
 @property double animationTime;
-@property int animationStage, tag;
+@property NSInteger animationStage, tag;
 @property id<AnimationManagerDelegate> delegate;
 @property NSArray *animationObjects;
 @property NSTimer *timer;
